[UPDATE] copy 명령어 중 폴더를 copy할경우 폴더 내부의 파일들이 copy 되도록 개선

+++ cmd/src/controller/command/Copy.java
@@ -7,6 +7,7 @@ import java.nio.file.StandardCopyOption;
 
 import controller.CmdService;
 import model.UserPath;
+import utility.Constant;
 import utility.DataProcessing;
 import view.CmdView;
 
@@ -28,8 +29,11 @@ public class Copy extends Move implements CmdService
 			return;
 		}
 		beforePath = getBeforePath(inputCommand, getBeforeFileName(inputCommand)) + getBeforeFileName(inputCommand);
-		afterPath = getAfterPath(inputCommand, getAfterFileName(inputCommand)) + getAfterFileName(inputCommand);
-		if (DataProcessing.get().isValidPath(beforePath))
+		afterPath = getAfterPath(inputCommand, getAfterFileName(inputCommand)) + getAfterFileName(inputCommand);		
+		File beforeFile = new File(beforePath);
+		if (DataProcessing.get().isValidPath(beforePath) && beforeFile.isDirectory())
+			copyDirectory(beforePath, afterPath);
+		else if (DataProcessing.get().isValidPath(beforePath) && beforeFile.isFile())
 			copyFile(beforePath, afterPath);
 		else
 			cmdView.print("지정된 파일을 찾을 수 없습니다.\n");
@@ -46,12 +50,59 @@ public class Copy extends Move implements CmdService
 		{
 			if (e.toString().contains("FileAlreadyExistsException"))
 			{
-				if (isReplaceIfExistFile(afterPath))
+				if (isReplaceIfExistFile(afterPath) == Constant.ReplaceOption.YES.getIndex() || isReplaceIfExistFile(afterPath) == Constant.ReplaceOption.ALL.getIndex())
+				{
 					copyFileOnReplaceOption(beforePath, afterPath);
+					cmdView.print("\t1개 파일이 복사되었습니다.\n");
+				}
 				else
 					cmdView.print("\t0개 파일이 복사되었습니다.\n");
 			}
 		}
+	}
+	
+	private void copyDirectory(String beforePath, String afterPath)
+	{
+		boolean isALLOption = false;
+		int movedFileCount = 0, replaceOption;
+		File beforeFile = new File(beforePath);
+		File[] fileList = beforeFile.listFiles(DataProcessing.get().fileFilter);
+		for (File fileName : fileList)
+		{
+			if(beforePath.equals(afterPath))
+			{
+				afterPath = DataProcessing.get().moveUpPathStage(afterPath, 1);
+				afterPath = afterPath + getBeforeFileName(fileName.getPath().replace("\\","\\\\"));
+			}
+			cmdView.print(fileName.toString().replace(userPath.get() + "\\", "") + "\n");
+			try 
+			{
+				if (isALLOption)
+					copyFileOnReplaceOption(fileName.getPath(), afterPath);
+				else
+					Files.copy(fileName.toPath(), new File(afterPath).toPath());
+				movedFileCount++;
+			}
+			catch (IOException e)
+			{
+				if (e.toString().contains("FileAlreadyExistsException"))
+				{
+					replaceOption = isReplaceIfExistFile(afterPath);
+					if (replaceOption == Constant.ReplaceOption.YES.getIndex())
+					{
+						copyFileOnReplaceOption(fileName.getPath(), afterPath);
+						movedFileCount++;
+					}
+					else if (replaceOption == Constant.ReplaceOption.ALL.getIndex())
+					{
+						copyFileOnReplaceOption(fileName.getPath(), afterPath);
+						movedFileCount++;
+						isALLOption = true;
+					}
+				}
+			}
+		}
+		cmdView.print("\t" + movedFileCount + "개 파일이 복사되었습니다.\n");
 	}
 	
 	private void copyFileOnReplaceOption(String beforePath, String afterPath)
@@ -59,7 +110,6 @@ public class Copy extends Move implements CmdService
 		try
 		{
 			Files.copy(new File(beforePath).toPath(), new File(afterPath).toPath(), StandardCopyOption.REPLACE_EXISTING);
-			cmdView.print("\t1개 파일이 복사되었습니다.\n");
 		}
 		catch (IOException e)
 		{
